Arduino Code
The purpose of the Arduino's code (written in C) is to take data from a
previously generated file and send each value out 2 bytes at a time as
voltage levels\footnote{While normally, values can be sent out in one
byte, in this case, one byte is not specific enough to properly
represent the needed value at such a large range of numbers} to the
electronic device that is being measured. Then it reads the voltage
output from the electronic device and sends it to the computer over the
serial port. (Add why timing is an issue and why we are limited to a
sample rate of 2). The program runs on a timer interrupt whose rate is
set by the sample rate. Each time the interrupt is run, it dumps two
words to the codec through a serial link, and it receives two words from
the codec. The it takes the two bytes from the left channel and writes
it to the serial usb link one byte at a time. After every two bytes, we
write a line feed character in order to make synchronization on the
receiving end easier. The serial link's baud rate is 115200, which is
limiting the sample rate that can be used.
We have also implemented a checksum to check for data integrity. This is
done every 128 words, where it takes the least sugnificant word of the
sum of the previous 128 data words, where it is later checked via the
Processing code. We identify the checksum with the dle(0x0020) character
followed by a two(0x0002), so then the Processing code knows to look for
a checksum.The purpose of the Arduino's code (written in C) is to take
data from a previously generated file and send each value out 2 bytes at
a time as voltage levels\footnote{While normally, values can be sent out
in one byte, in this case, one byte is not specific enough to properly
represent the needed value at such a large range of numbers} to the
electronic device that is being measured. Then it reads the voltage
output from the electronic device and sends it to the computer over the
serial port. (Add why timing is an issue and why we are limited to a
sample rate of 2). The program runs on a timer interrupt whose rate is
set by the sample rate. Each time the interrupt is run, it dumps two
words to the codec through a serial link, and it receives two words from
the codec. The it takes the two bytes from the left channel and writes
it to the serial usb link one byte at a time. After every two bytes, we
write a line feed character in order to make synchronization on the
receiving end easier. The serial link's baud rate is 115200, which is
limiting the sample rate that can be used.
We have also implemented a checksum to check for data integrity. This is
done every 128 words, where it takes the least sugnificant word of the
sum of the previous 128 data words, where it is later checked via the
Processing code. We identify the checksum with the dle(0x0020) character
followed by a two(0x0002), so then the Processing code knows to look for
a checksum.