Processing Code
The purpose of the Processing code is to receive the data coming from
the Arduino and to organize it into a way that is easy to later be
interpreted by a separate program. The processing file graphs the data
points so the user can see if the data looks correct, and it also stores
the data in separate text files so they can later be analyzed. The
Processing code accepts bytes until it sees a line feed, which generates
an interrupt. The interrupt reads the first two bytes to determine if it
is a checksum(dle followed by two) or actual data. If it is data, it
gets accumulated into the checksum and written to an element of an array
which will then be plotted on the ongoing Processing window. However, if
the first two bytes were a dle-2, then the program compares the
subsequent two bytes with the local checksum, where if there is a
decrepancy, it is printed to the console as an error.
If we have reached a total of 1024 valid data words, we initiate a
thread to write the data to the disk, The data writing thread is set up
to write ten text files of 1024 words named dataXX.txt, where XX is 0 to
9. Subsequent data is not written.