iPython Notebook Code
The purpose of the IPython code is to read the input and output data and
calculate the transfer function from the input to the output, and then
graphs the transfer function so the student can visualize how each
frequency is transformed. We define a directory which contains all the
data from the Processing sketches, and it reads and averages all the
data in all the files.  Next we input a copy of the same bandlimited
file that had been sent to the Arduino code and it plots that input code
and the resultant data on top of each other. Then we fourier transform
the original and resulting data, and take the ratio of the two, where we
graph it along side the ideal model. Finally, we make a plot on a log
log axis of the transfer function as a function of frequency. Lastly, we
plot the residual of the data compared to the model.